pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Natalija <vukosavljenatalija@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.30"
databases = "^0.9.0"
alembic = "^1.13.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
python-jose = "^3.3.0"
bcrypt = "4.0.1"
requests = "^2.32.3"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
boto3 = "^1.34.138"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"

[tool.ruff]
# Set the maximum line length to 79.
line-length = 79

[tool.ruff.lint]
select = ["ALL"]



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
